Allow shadow to run jobs without a schedd feeding it. #1374
This is a first cut, there is more work to be done, but this
at least gets us started.

+++ src/condor_c++_util/condor_claimid_parser.h
@@ -127,7 +127,7 @@ class ClaimIdParser {
 			if(!endptr || endptr < ptr) {
 				return NULL;
 			}
-			m_session_info.sprintf("%.*s",endptr+1-ptr,ptr);
+			m_session_info.sprintf("%.*s",(int)(endptr+1-ptr),ptr);
 		}
 
 		if( m_session_info.IsEmpty() ) {
